# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jastip Anggun Jaya API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for customer, admin, or owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers (admin/owner)
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerWithStats"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerWithStats"

  # ─── ADMINS ─────────────────────────────────────────────────────────────────
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List all admins with package stats (owner only)
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminWithStats"

    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create a new admin (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admins/{id}:
    patch:
      operationId: updateAdmin
      tags: [admins]
      summary: Update admin (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdate"
      responses:
        "200":
          description: Admin updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admins/{id}/toggle-active:
    post:
      operationId: toggleAdminActive
      tags: [admins]
      summary: Activate or deactivate admin (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admins/{id}/reset-password:
    post:
      operationId: resetAdminPassword
      tags: [admins]
      summary: Reset admin password (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
  /transactions/{id}/receipt:
    get:
      operationId: getTransactionReceipt
      tags: [transactions]
      summary: Get receipt data for a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionReceipt"
        "404":
          description: Transaction not found
  /transactions/{id}/receipt/print:
    post:
      operationId: printTransactionReceipt
      tags: [transactions]
      summary: Record a transaction receipt print
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt data and print log information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintedTransactionReceipt"
        "404":
          description: Transaction not found
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    TransactionReceipt:
      type: object
      required: [business, transaction, payments, packages]
      properties:
        business:
          type: object
          required: [name, subtitle]
          properties:
            name:
              type: string
            subtitle:
              type: string
        transaction:
          $ref: "#/components/schemas/ReceiptTransaction"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptPayment"
        packages:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptPackage"
        cashier:
          anyOf:
            - $ref: "#/components/schemas/ReceiptCashier"
            - type: "null"
        shift:
          anyOf:
            - $ref: "#/components/schemas/ReceiptShift"
            - type: "null"
    PrintedTransactionReceipt:
      type: object
      required: [receipt, print]
      properties:
        receipt:
          $ref: "#/components/schemas/TransactionReceipt"
        print:
          type: object
          required: [copyNumber, isReprint]
          properties:
            copyNumber:
              type: integer
            isReprint:
              type: boolean
            label:
              type: ["string", "null"]
    ReceiptTransaction:
      type: object
      required: [id, transactionNo, customerName, subtotal, discount, total, paymentStatus, transactionStatus, sisaPiutang, createdAt, packageIds]
      properties:
        id:
          type: integer
        transactionNo:
          type: string
        customerName:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        discountReason:
          type: ["string", "null"]
        total:
          type: number
        paymentStatus:
          type: string
        transactionStatus:
          type: string
        sisaPiutang:
          type: number
        createdAt:
          type: string
        packageIds:
          type: array
          items:
            type: integer
    ReceiptPayment:
      type: object
      properties:
        id:
          type: integer
        paymentType:
          type: string
        paymentTypeLabel:
          type: string
        paymentMethod:
          type: ["string", "null"]
        paymentMethodLabel:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        changeAmount:
          type: number
        paymentReference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ReceiptPackage:
      type: object
      properties:
        id:
          type: integer
        barcode:
          type: string
        resiNumber:
          type: string
        packageNumber:
          type: ["string", "null"]
        itemName:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        deliveryRoute:
          type: ["string", "null"]
        usedWeight:
          type: number
        totalShipping:
          type: number
    ReceiptCashier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
    ReceiptShift:
      type: object
      properties:
        id:
          type: integer
        shiftType:
          type: string
        terminalId:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, phone, role, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, owner]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    CustomerWithStats:
      type: object
      required:
        [
          id,
          name,
          phone,
          isActive,
          totalPackages,
          pendingPackages,
          pickedUpPackages,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        totalPackages:
          type: integer
        pendingPackages:
          type: integer
        pickedUpPackages:
          type: integer
    AdminWithStats:
      type: object
      required: [id, name, phone, isActive, packagesInputted]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        packagesInputted:
          type: integer
    AdminInput:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
    AdminUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
...